% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Chr}
\alias{Chr}
\title{Character scalar type constructor}
\arguments{
\item{.x}{A character vector}
}
\value{
An object with type \code{Chr}
}
\description{
Character scalar type constructor
}
\examples{
Chr("a")
}
