% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Chr_}
\alias{Chr_}
\title{Character vector type constructor}
\arguments{
\item{.x}{A character vector}
}
\value{
An object with type \code{Chr_}
}
\description{
Character vector type constructor
}
\examples{
Chr_(c("a", "b"))
}
