% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Dbl_}
\alias{Dbl_}
\title{Double vector type constructor}
\arguments{
\item{.x}{A double vector}
}
\value{
An object with type \code{Dbl_}
}
\description{
Double vector type constructor
}
\examples{
Dbl_(c(1, 2))
}
