% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==ss\%}
\alias{\%==ss\%}
\title{Check equality of two scalars}
\arguments{
\item{.x}{A scalar}

\item{.y}{A scalar}
}
\value{
A logical scalar
}
\description{
Check equality of two scalars
}
\examples{
Int(1L) \%==ss\% Int(1L)
}
\seealso{
Other equal functions: 
\code{\link{\%==sv\%}()},
\code{\link{\%==vs\%}()},
\code{\link{\%==vv\%}()},
\code{\link{pure_equal}()}
}
\concept{equal functions}
