% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==sv\%}
\alias{\%==sv\%}
\title{Check equality of a scalar and vector}
\arguments{
\item{.x}{A scalar}

\item{.y}{A vector}
}
\value{
A logical vector
}
\description{
Check equality of a scalar and vector
}
\examples{
Int(1L) \%==ss\% Int_(c(1L, 2L))
}
\seealso{
Other equal functions: 
\code{\link{\%==ss\%}()},
\code{\link{\%==vs\%}()},
\code{\link{\%==vv\%}()},
\code{\link{pure_equal}()}
}
\concept{equal functions}
