% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==vs\%}
\alias{\%==vs\%}
\title{Check equality of a vector and scalar}
\arguments{
\item{.x}{A vector}

\item{.y}{A scalar}
}
\value{
A logical vector
}
\description{
Check equality of a vector and scalar
}
\examples{
Int_(c(1L, 2L)) \%==ss\% Int(1L)
}
\seealso{
Other equal functions: 
\code{\link{\%==ss\%}()},
\code{\link{\%==sv\%}()},
\code{\link{\%==vv\%}()},
\code{\link{pure_equal}()}
}
\concept{equal functions}
