% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==vv\%}
\alias{\%==vv\%}
\title{Check equality of two vectors}
\arguments{
\item{.x}{A vector}

\item{.y}{A vector}
}
\value{
A logical vector
}
\description{
Check equality of two vectors
}
\examples{
Int_(c(1L, 2L)) \%==ss\% Int_(c(1L, 2L))
}
\seealso{
Other equal functions: 
\code{\link{\%==ss\%}()},
\code{\link{\%==sv\%}()},
\code{\link{\%==vs\%}()},
\code{\link{pure_equal}()}
}
\concept{equal functions}
