% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%>=sv\%}
\alias{\%>=sv\%}
\title{Greater than or equal scalar to vector number comparison}
\arguments{
\item{.x}{A number scalar}

\item{.y}{A number vector}
}
\value{
A logical vector
}
\description{
Greater than or equal scalar to vector number comparison
}
\examples{
Int(1L) \%>=sv\% Int_(c(1L, 2L))
}
\seealso{
Other gte functions: 
\code{\link{\%>=ss\%}()},
\code{\link{\%>=vs\%}()},
\code{\link{\%>=vv\%}()},
\code{\link{pure_gte}()}
}
\concept{gte functions}
