% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%>=vv\%}
\alias{\%>=vv\%}
\title{Greater than or equal vector number comparison}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number vector}
}
\value{
A logical vector
}
\description{
Greater than or equal vector number comparison
}
\examples{
Int_(c(1L, 2L)) \%>=vv\% Int_(c(1L, 1L))
}
\seealso{
Other gte functions: 
\code{\link{\%>=ss\%}()},
\code{\link{\%>=sv\%}()},
\code{\link{\%>=vs\%}()},
\code{\link{pure_gte}()}
}
\concept{gte functions}
