% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%<=ss\%}
\alias{\%<=ss\%}
\title{Less than or equal scalar number comparison}
\arguments{
\item{.x}{A number scalar}

\item{.y}{A number scalar}
}
\value{
A logical scalar
}
\description{
Less than or equal scalar number comparison
}
\examples{
Int(1L) \%<=ss\% Int(1L)
}
\seealso{
Other lte functions: 
\code{\link{\%<=sv\%}()},
\code{\link{\%<=vs\%}()},
\code{\link{\%<=vv\%}()},
\code{\link{pure_lte}()}
}
\concept{lte functions}
