% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%<=vs\%}
\alias{\%<=vs\%}
\title{Less than or equal vector to scalar number comparison}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number scalar}
}
\value{
A logical vector
}
\description{
Less than or equal vector to scalar number comparison
}
\examples{
Int_(c(1L, 2L)) \%<=vs\% Int(1L)
}
\seealso{
Other lte functions: 
\code{\link{\%<=ss\%}()},
\code{\link{\%<=sv\%}()},
\code{\link{\%<=vv\%}()},
\code{\link{pure_lte}()}
}
\concept{lte functions}
