% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%modss\%}
\alias{\%modss\%}
\title{Get modulo of two scalar numbers}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar with the same type as input
}
\description{
Get modulo of two scalar numbers
}
\examples{
Dbl(1) \%modss\% Dbl(1)
Int(1L) \%modss\% Int(1L)
}
\seealso{
Other modulo functions: 
\code{\link{\%modvs\%}()},
\code{\link{\%modvv\%}()},
\code{\link{pure_modulo}()}
}
\concept{modulo functions}
