% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%modsv\%}
\alias{\%modsv\%}
\title{Get modulo of a number scalar and a number vector}
\arguments{
\item{.x}{A number scalar}

\item{.y}{A number vector}
}
\value{
A vector with the same type as input
}
\description{
Get modulo of a number scalar and a number vector
}
\examples{
Dbl(1) \%modsv\% Dbl_(c(1, 2))
Int(1) \%modsv\% Int_(c(1L, 2L))
}
\seealso{
Other add functions: 
\code{\link{\%+ss\%}()},
\code{\link{\%+sv\%}()},
\code{\link{\%+vs\%}()},
\code{\link{\%+vv\%}()},
\code{\link{pure_add}()}
}
\concept{add functions}
