% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%modvs\%}
\alias{\%modvs\%}
\title{Get modulo of a number vector and a number scalar}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number scalar}
}
\value{
A vector with the same type as input
}
\description{
Get modulo of a number vector and a number scalar
}
\examples{
Dbl_(c(1, 2)) \%modvs\% Dbl(1)
Int_(c(1L, 2L)) \%modvs\% Int(1L)
}
\seealso{
Other modulo functions: 
\code{\link{\%modss\%}()},
\code{\link{\%modvv\%}()},
\code{\link{pure_modulo}()}
}
\concept{modulo functions}
