% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%modvv\%}
\alias{\%modvv\%}
\title{Get modulo of two number vectors}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number vector}
}
\value{
A vector with the same type as input
}
\description{
Get modulo of two number vectors
}
\examples{
Dbl_(c(1, 2)) \%modvv\% Dbl_(c(1, 2))
Int_(c(1L, 2L)) \%modvv\% Int_(c(1L, 2L))
}
\seealso{
Other modulo functions: 
\code{\link{\%modss\%}()},
\code{\link{\%modvs\%}()},
\code{\link{pure_modulo}()}
}
\concept{modulo functions}
