% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+sv\%}
\alias{\%+sv\%}
\title{Add a number scalar to a number vector}
\arguments{
\item{.x}{A number scalar}

\item{.y}{A number vector}
}
\value{
A vector with the same type as input
}
\description{
Add a number scalar to a number vector
}
\examples{
Dbl(1) \%+sv\% Dbl_(c(1, 2))
Int(1) \%+sv\% Int_(c(1L, 2L))
}
\seealso{
Other add functions: 
\code{\link{\%+ss\%}()},
\code{\link{\%+vs\%}()},
\code{\link{\%+vv\%}()},
\code{\link{\%modsv\%}()},
\code{\link{pure_add}()}
}
\concept{add functions}
