% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+vs\%}
\alias{\%+vs\%}
\title{Add a number vector to a number scalar}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number scalar}
}
\value{
A vector with the same type as input
}
\description{
Add a number vector to a number scalar
}
\examples{
Dbl_(c(1, 2)) \%+vs\% Dbl(1)
Int_(c(1L, 2L)) \%+vs\% Int(1L)
}
\seealso{
Other add functions: 
\code{\link{\%+ss\%}()},
\code{\link{\%+sv\%}()},
\code{\link{\%+vv\%}()},
\code{\link{\%modsv\%}()},
\code{\link{pure_add}()}
}
\concept{add functions}
