% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+vv\%}
\alias{\%+vv\%}
\title{Add two number vectors}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number vector}
}
\value{
A vector with the same type as input
}
\description{
Add two number vectors
}
\examples{
Dbl_(c(1, 2)) \%+vv\% Dbl_(c(1, 2))
Int_(c(1L, 2L)) \%+vv\% Int_(c(1L, 2L))
}
\seealso{
Other add functions: 
\code{\link{\%+ss\%}()},
\code{\link{\%+sv\%}()},
\code{\link{\%+vs\%}()},
\code{\link{\%modsv\%}()},
\code{\link{pure_add}()}
}
\concept{add functions}
