% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%/sv\%}
\alias{\%/sv\%}
\title{Divide a number scalar by a number vector}
\arguments{
\item{.x}{A number scalar}

\item{.y}{A number vector}
}
\value{
A vector with the same kind as inputs
}
\description{
Divide a number scalar by a number vector
}
\examples{
Dbl(1) \%/sv\% Dbl_(c(1, 2))
Int(1) \%/sv\% Int_(c(1, 2))
}
\seealso{
Other divide functions: 
\code{\link{\%/ss\%}()},
\code{\link{\%/vs\%}()},
\code{\link{\%/vv\%}()},
\code{\link{pure_divide}()}
}
\concept{divide functions}
