% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%/vv\%}
\alias{\%/vv\%}
\title{Divide a number vector by a number vector}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number vector}
}
\value{
A vector with the same kind as inputs
}
\description{
Divide a number vector by a number vector
}
\examples{
Dbl_(c(1, 2)) \%/vv\% Dbl_(c(1, 2))
Int_(c(1, 2)) \%/vv\% Int_(c(1, 2))
}
\seealso{
Other divide functions: 
\code{\link{\%/ss\%}()},
\code{\link{\%/sv\%}()},
\code{\link{\%/vs\%}()},
\code{\link{pure_divide}()}
}
\concept{divide functions}
