% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%-ss\%}
\alias{\%-ss\%}
\title{Subtract two scalar numbers}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar with the same type as input
}
\description{
Subtract two scalar numbers
}
\examples{
Dbl(1) \%-ss\% Dbl(1)
Int(1L) \%-ss\% Int(1L)
}
\seealso{
Other subtract functions: 
\code{\link{\%-sv\%}()},
\code{\link{\%-vs\%}()},
\code{\link{\%-vv\%}()},
\code{\link{pure_subtract}()}
}
\concept{subtract functions}
