% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%*vv\%}
\alias{\%*vv\%}
\title{Multiply two number vectors}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number vector}
}
\value{
A vector with the same type as input
}
\description{
Multiply two number vectors
}
\examples{
Dbl(1) \%*vv\% Dbl(1)
Int(1L) \%*vv\% Int(1L)
}
\seealso{
Other multiply functions: 
\code{\link{\%*ss\%}()},
\code{\link{\%*sv\%}()},
\code{\link{\%*vs\%}()},
\code{\link{pure_multiply}()}
}
\concept{multiply functions}
