% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%-vs\%}
\alias{\%-vs\%}
\title{Subtract a number scalar from a number vector}
\arguments{
\item{.x}{A number vector}

\item{.y}{A number scalar}
}
\value{
A vector with the same type as input
}
\description{
Subtract a number scalar from a number vector
}
\examples{
Dbl_(c(1, 2)) \%-vs\% Dbl(1)
Int_(c(1L, 2L)) \%-vs\% Int(1L)
}
\seealso{
Other subtract functions: 
\code{\link{\%-ss\%}()},
\code{\link{\%-sv\%}()},
\code{\link{\%-vv\%}()},
\code{\link{pure_subtract}()}
}
\concept{subtract functions}
