% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Int}
\alias{Int}
\title{Integer scalar type constructor}
\arguments{
\item{.x}{An integer scalar}
}
\value{
An object with type \code{Int}
}
\description{
Integer scalar type constructor
}
\examples{
Int(1L)
}
