% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Int_}
\alias{Int_}
\title{Integer vector type constructor}
\arguments{
\item{.x}{An integer vector}
}
\value{
An object with type \code{Int_}
}
\description{
Integer vector type constructor
}
\examples{
Int_(1L)
}
