% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Lgl}
\alias{Lgl}
\title{Logical scalar type constructor}
\arguments{
\item{.x}{A logical vector}
}
\value{
An object with type \code{Lgl}
}
\description{
Logical scalar type constructor
}
\examples{
Lgl("a")
}
