% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Lgl_}
\alias{Lgl_}
\title{Logical vector type constructor}
\arguments{
\item{.x}{A logical vector}
}
\value{
An object with type \code{Lgl_}
}
\description{
Logical vector type constructor
}
\examples{
Lgl_(c(TRUE, FALSE))
}
