% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_add}
\alias{pure_add}
\alias{add}
\title{Add two numbers together}
\usage{
add(...)
}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Add two numbers together
}
\examples{
pure_add(Int(1L), Int(1L), .types = Chr("ss"))
pure_add(Dbl(1), Dbl_(c(1, 2)), .types = Chr("sv"))
}
\seealso{
Other add functions: 
\code{\link{\%+ss\%}()},
\code{\link{\%+sv\%}()},
\code{\link{\%+vs\%}()},
\code{\link{\%+vv\%}()},
\code{\link{\%modsv\%}()}
}
\concept{add functions}
