% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_divide}
\alias{pure_divide}
\alias{divi}
\title{Divide two numbers}
\usage{
divi(...)
}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Divide two numbers
}
\examples{
pure_divide(Int(1L), Int(1L), .types = Chr("ss"))
pure_divide(Dbl(1), Dbl_(c(1, 2)), .types = Chr("sv"))
}
\seealso{
Other divide functions: 
\code{\link{\%/ss\%}()},
\code{\link{\%/sv\%}()},
\code{\link{\%/vs\%}()},
\code{\link{\%/vv\%}()}
}
\concept{divide functions}
