% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_equal}
\alias{pure_equal}
\alias{equal}
\title{Check equality of two scalars or vectors}
\usage{
equal(...)
}
\arguments{
\item{.x}{A scalar or vector}

\item{.y}{A scalar or vector}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Check equality of two scalars or vectors
}
\examples{
pure_equal(Int(1L), Int(1L), .types = Chr("ss"))
pure_equal(Chr("a"), Chr_(c("a", "b")), .types = Chr("sv"))
}
\seealso{
Other equal functions: 
\code{\link{\%==ss\%}()},
\code{\link{\%==sv\%}()},
\code{\link{\%==vs\%}()},
\code{\link{\%==vv\%}()}
}
\concept{equal functions}
