% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_gt}
\alias{pure_gt}
\alias{gt}
\title{Greater than number comparison}
\usage{
gt(...)
}
\arguments{
\item{.x}{A scalar or vector}

\item{.y}{A scalar or vector}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Greater than number comparison
}
\examples{
pure_gt(Int(1L), Int(1L), .types = Chr("ss"))
pure_gt(Dbl(1), Dbl(1), .types = Chr("ss"))
}
\seealso{
Other gt functions: 
\code{\link{\%>ss\%}()},
\code{\link{\%>sv\%}()},
\code{\link{\%>vs\%}()},
\code{\link{\%>vv\%}()}
}
\concept{gt functions}
