% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_gte}
\alias{pure_gte}
\alias{gte}
\title{Greater than or equal number comparison}
\usage{
gte(...)
}
\arguments{
\item{.x}{A scalar or vector}

\item{.y}{A scalar or vector}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Greater than or equal number comparison
}
\examples{
pure_gte(Int(1L), Int(1L), .types = Chr("ss"))
pure_gte(Dbl(1), Dbl(1), .types = Chr("ss"))
}
\seealso{
Other gte functions: 
\code{\link{\%>=ss\%}()},
\code{\link{\%>=sv\%}()},
\code{\link{\%>=vs\%}()},
\code{\link{\%>=vv\%}()}
}
\concept{gte functions}
