% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_lt}
\alias{pure_lt}
\alias{lt}
\title{Less than number comparison}
\usage{
lt(...)
}
\arguments{
\item{.x}{A scalar or vector}

\item{.y}{A scalar or vector}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Less than number comparison
}
\examples{
pure_lt(Int(1L), Int(1L), .types = Chr("ss"))
pure_lt(Dbl(1), Dbl(1), .types = Chr("ss"))
}
\seealso{
Other lt functions: 
\code{\link{\%<ss\%}()},
\code{\link{\%<sv\%}()},
\code{\link{\%<vs\%}()},
\code{\link{\%<vv\%}()}
}
\concept{lt functions}
