% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_lte}
\alias{pure_lte}
\alias{lte}
\title{Less than number comparison}
\usage{
lte(...)
}
\arguments{
\item{.x}{A scalar or vector}

\item{.y}{A scalar or vector}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Less than number comparison
}
\examples{
pure_lte(Int(1L), Int(1L), .types = Chr("ss"))
pure_lte(Dbl(1), Dbl(1), .types = Chr("ss"))
}
\seealso{
Other lte functions: 
\code{\link{\%<=ss\%}()},
\code{\link{\%<=sv\%}()},
\code{\link{\%<=vs\%}()},
\code{\link{\%<=vv\%}()}
}
\concept{lte functions}
