% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_modulo}
\alias{pure_modulo}
\alias{modu}
\title{Get modulo of two numbers}
\usage{
modu(...)
}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Get modulo of two numbers
}
\examples{
pure_modulo(Int(1L), Int(1L), .types = Chr("ss"))
pure_modulo(Dbl(1), Dbl_(c(1, 2)), .types = Chr("sv"))
}
\seealso{
Other modulo functions: 
\code{\link{\%modss\%}()},
\code{\link{\%modvs\%}()},
\code{\link{\%modvv\%}()}
}
\concept{modulo functions}
