% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_multiply}
\alias{pure_multiply}
\alias{mult}
\title{Multiply two numbers}
\usage{
mult(...)
}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Multiply two numbers
}
\examples{
pure_multiply(Int(1L), Int(1L), .types = Chr("ss"))
pure_multiply(Dbl(1), Dbl_(c(1, 2)), .types = Chr("sv"))
}
\seealso{
Other multiply functions: 
\code{\link{\%*ss\%}()},
\code{\link{\%*sv\%}()},
\code{\link{\%*vs\%}()},
\code{\link{\%*vv\%}()}
}
\concept{multiply functions}
