% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pure_subtract}
\alias{pure_subtract}
\alias{subt}
\title{Subtract two numbers}
\usage{
subt(...)
}
\arguments{
\item{.x}{A number}

\item{.y}{A number}
}
\value{
A scalar or a vector depending on \code{.types}
}
\description{
Subtract two numbers
}
\examples{
pure_subtract(Int(1L), Int(1L), .types = Chr("ss"))
pure_subtract(Dbl(1), Dbl_(c(1, 2)), .types = Chr("sv"))
}
\seealso{
Other subtract functions: 
\code{\link{\%-ss\%}()},
\code{\link{\%-sv\%}()},
\code{\link{\%-vs\%}()},
\code{\link{\%-vv\%}()}
}
\concept{subtract functions}
