% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Tbl}
\alias{Tbl}
\title{Pure tibble type constructor}
\arguments{
\item{.x}{A tibble}

\item{.types}{A list of column types.
List names are the column names of the tibble
and the list values are the pure vector type.}
}
\value{
An object with type \code{Tbl}
}
\description{
Pure tibble type constructor
}
\examples{
Tbl(
  tibble::tibble(
    a = c(1, 2),
    b = c("a", "b")
  ),
  .types = list(
    a = "Dbl_",
    b = "Chr_"
  )
)
}
